Removed extra class and moved it within LibraryBookFatima
